added decomposer structure. pushed rand version
