+++ Cargo.toml
@@ -19,7 +19,7 @@ itertools = "0.13.0"
 ndarray = "0.16.1"
 openqasm = "0.1.2"
 pyo3 = { version = "0.25.1" }
-rand = "0.8.3"
+rand = { version = "0.9.2", features = ["small_rng"] }
 rayon = "1.5.0"
 regex = "1.11.1"
 rustc-hash = "2.1.1"
